requirements_new.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
fastapi==0.129.0
ffmpy==1.0.0
filelock==3.24.3
fsspec==2026.2.0
gradio==6.6.0
gradio-client==2.1.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
pyyaml==6.0.3
regex==2026.2.19
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.0
typer-slim==0.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
